+++ backend/services/auth-service-backend/src/main/java/br/com/condominium/auth/login/controller/AuthController.java
@@ -1,7 +1,7 @@
 package br.com.condominium.auth.login.controller;
 
-import br.com.condominium.auth.jwt.dto.LoginInput;
-import br.com.condominium.auth.jwt.dto.LoginOutput;
+import br.com.condominium.auth.login.dto.LoginInput;
+import br.com.condominium.auth.login.dto.LoginOutput;
 import br.com.condominium.auth.login.service.AuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;