+++ backend/services/auth-service-backend/src/main/java/br/com/condominium/auth/login/dto/LoginOutput.java
@@ -1,4 +1,5 @@
 package br.com.condominium.auth.login.dto;
 
-public class LoginOutput {
-}
+public record LoginOutput(
+        String accessToken
+) {}